fix(pi0): cast SigLIP output to fp32 at the boundary
DepthExpert (fp32) hit LayerNorm dtype mismatch with bf16 SigLIP
features. Upcast at the encode_image_to_siglip boundary keeps SigLIP
itself in bf16 (memory + speed) while letting the downstream
DepthExpert + VQVAE codebook stay fp32 without per-module dtype
shenanigans.

+++ src/georel_vla/backbones/pi0.py
@@ -202,7 +202,10 @@ class Pi0Backbone(VLABackbone):
                 f"SigLIP output shape {tuple(out.shape)} does not match "
                 f"(B, {self.n_image_tokens}, {self.siglip_dim})"
             )
-        return out
+        # Cast back to float32 so the downstream DepthExpert (which is fp32 by
+        # default) doesn't trip LayerNorm's dtype check. SigLIP itself ran in
+        # bf16 (memory + speed); the boundary upcast is cheap.
+        return out.float()
 
     def _init_processor(self) -> None:
         """Lazy-init the VLAProcessor + PaliGemma tokenizer."""
